Added remaining DAO and Services for all Models. Added admin pages and autogenerated admin. Implemented admin create account.

+++ src/main/java/com/photocontest/model/Voter.java
@@ -19,5 +19,5 @@ public class Voter implements Serializable {
     @Column(name = "voter_id")
     private long voter_id;
 
-    private int ip_address;
+    private long ip_address;
 }
